Fix warning (circular dep)
