+++ notifications/index.js
@@ -8,6 +8,7 @@ const {
   trySendPushover,
   PUSHOVER_ENABLED
 } = require('./pushover');
+const { getVersionMessage } = require('../utils');
 
 const lastSentMessages = new Map();
 
@@ -46,7 +47,26 @@ const initNotifications = async () => {
   await sendNotifications(`Starting up sat hunter on ${process.env.ACTIVE_EXCHANGE}`)
 };
 
+/**
+ * Sends notifications about the version message
+ * @param interval Timeout in milliseconds, default is 24 hours
+ * @returns {Promise<void>}
+ */
+const initVersionCheck = async (interval = 1000 * 60 * 60 * 24) => {
+    console.log('Checking version...')
+    console.log(await getVersionMessage());
+
+    const runPeriodicVersionCheck = async () => {
+        const message = await getVersionMessage();
+        await sendNotifications(message, 'version_check');
+    }
+
+    await runPeriodicVersionCheck();
+    setInterval(runPeriodicVersionCheck, interval);
+};
+
 module.exports = {
-  sendNotifications,
-  initNotifications
+    initVersionCheck,
+    sendNotifications,
+    initNotifications
 };
